+++ index.js
@@ -3,15 +3,14 @@
 
 const path = require('path');
 
-const {summarize} = require('./lib/summarize');
+const { summarize } = require('./lib/summarize');
+const makeBenchmark = require('./lib/make-benchmark');
 
 /* eslint-disable no-console */
 /* eslint-disable no-unused-vars */
 
 // this file executes tests imported from a definitions file.
 
-const perf_hooks = require('perf_hooks');
-const {performance: perf, PerformanceObserver: PerfObserver} = perf_hooks;
 const util = require('util');
 
 
@@ -23,6 +22,7 @@ let benchmarkFile = './benchmarks/definitions';
 if (process.env.BENCH) {
   benchmarkFile = path.resolve(process.env.BENCH);
 }
+const verbose = process.env.VERBOSE;
 const definitions = require(benchmarkFile);
 
 const {
@@ -39,10 +39,10 @@ if (!tests.noop) {
 }
 // noop that returns undefined
 if (!tests.noopUndef) {
-  tests.noopUndef = async() => undefined;
+  tests.noopUndef = async () => undefined;
 }
 if (!tests.noopu) {
-  tests.noopu = async() => undefined;
+  tests.noopu = async () => undefined;
 }
 if (!tests.debug) {
   tests.debug = async s => {
@@ -99,19 +99,15 @@ const {
   stddevRange,
 } = config;
 
-if (!stddevRange) {
-  console.log('the stddevRange must be a non-zero number:', stddevRange);
+if (!stddevRange || stddevRange <= 0 || typeof stddevRange !== 'number') {
+  console.log('the stddevRange must be a positive number:', stddevRange);
   process.exit(1);
 }
 //
 // this mess of args parsing probably belongs in another module.
 //
-const groupTimes = [];
-let gcCounts = 0;
-let totalGCTime = 0;
 let memCheck = false;
 let debug = false;
-const memory = Array(groupCount);
 const functionNames = [];
 // take from env?
 let args = process.argv.slice(2);
@@ -153,7 +149,6 @@ for (const arg of args) {
     process.exit(1);
   }
 }
-config.functionChain = functionNames.slice();
 
 if (!json) {
   console.log(`[function chain: ${functionNames.join(', ')}]`);
@@ -172,120 +167,138 @@ if (!json) {
   }));
 }
 
-let gcTypes;
-const k = perf_hooks.constants;
-// major changed from 2 to 4 for some reason.
-if (k.NODE_PERFORMANCE_GC_MAJOR === 4) {
-  gcTypes = {
-    [k.NODE_PERFORMANCE_GC_MAJOR] : 'major',
-    [k.NODE_PERFORMANCE_GC_MINOR] : 'minor',
-    [k.NODE_PERFORMANCE_GC_INCREMENTAL] : 'incr',
-    [k.NODE_PERFORMANCE_GC_WEAKCB] : 'weak',
-  };
-} else {
-  gcTypes = {
-    [k.NODE_PERFORMANCE_GC_MINOR]: 'minor',      // 1
-    [k.NODE_PERFORMANCE_GC_MAJOR]: 'major',      // 2
-    [k.NODE_PERFORMANCE_GC_INCREMENTAL]: 'incr', // 4
-    [k.NODE_PERFORMANCE_GC_WEAKCB]: 'weak',      // 8
-  };
-}
-
-//
-// setup measurements with performance hooks
-//
-const verbose = process.env.VERBOSE;
-
-const obs = new PerfObserver((list) => {
-  const entries = list.getEntries();
-  for (let i = 0; i < entries.length; i++) {
-    const entry = entries[i];
-    if (entry.entryType === 'measure') {
-      verbose && console.log(`perf ${entry.name}: ${entry.duration}`);
-      groupTimes.push(entry.duration);
-    } else if (entry.entryType === 'gc') {
-      if (verbose) {
-        const {kind, flags} = entry.detail ? entry.detail : entry;
-        console.log(`perf gc: ${entry.duration} (${gcTypes[kind]}) flags: ${flags}`);
-      }
-      gcCounts += 1;
-      totalGCTime += entry.duration;
-    }
-  }
-});
-obs.observe({entryTypes: ['measure', 'gc'], buffered: true});
-
-//
-// function to run the setups, tests, and final.
-//
-async function test() {
-  // call the tester's setup
-  if (definitions.setup) {
-    await (async() => definitions.setup(config))();
-  }
-  if (groupSetup) {
-    await (async() => groupSetup(config))();
-  }
-  // warmup
-  for (let i = warmupIterations; i > 0; i--) {
-    await execute(functionChain);
-  }
-  await pause(groupWaitMS);
-
-  // execute x groups of y iterations with a pause after each group
-  for (let i = 0; i < groupCount; i++) {
-    // setup for the group
-    if (groupSetup) {
-      await (async() => groupSetup(config))();
-    }
-    perf.mark('start-iteration');
-    for (let i = groupIterations; i > 0; i--) {
-      await execute(functionChain);
-    }
-    perf.measure('iteration-time', 'start-iteration');
-    if (memCheck) {
-      memory[i] = process.memoryUsage();
-    }
-    await pause(groupWaitMS);
-  }
-
-  // final pause - wait extra time to let GC finish.
-  await pause(groupWaitMS * 10);
-  return pause(groupWaitMS);
-}
+const hooks = { setup, groupSetup, final };
+const runSettings = {
+  functionChain,
+  functionNames,
+  memCheck,
+  verbose,
+};
 
 
-//
-// execute functionChains
-//
-async function execute(fc) {
-  let lastResult = undefined;
-  for (let i = 0; i < fc.length; i++) {
-    lastResult = await fc[i](lastResult);
-  }
-}
+const options = { json, terse };
 
-async function pause(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
+const benchmark = makeBenchmark(config, hooks, runSettings);
 
-//
-// execute the tests then report
-//
-test().then(() => {
-  obs.disconnect();
-  const gTimes = groupTimes.slice();
-  const gcStats = {gcCounts, totalGCTime};
-  const data = {gTimes, gcStats, stddevRange, config};
-  if (memCheck) {
-    data.memory = memory;
-  }
-  // if the test-definitions specifies a final execute it. it can
-  // be output or cleanup or whatever.
-  if (final) {
-    final(config);
-  }
+benchmark()
+  .then(benchmarkData => {
+    summarize(benchmarkData, options);
+  });
 
-  summarize(data, { json, terse });
-});
+// let gcTypes;
+// const k = perf_hooks.constants;
+// // major changed from 2 to 4 for some reason.
+// if (k.NODE_PERFORMANCE_GC_MAJOR === 4) {
+//   gcTypes = {
+//     [k.NODE_PERFORMANCE_GC_MAJOR] : 'major',
+//     [k.NODE_PERFORMANCE_GC_MINOR] : 'minor',
+//     [k.NODE_PERFORMANCE_GC_INCREMENTAL] : 'incr',
+//     [k.NODE_PERFORMANCE_GC_WEAKCB] : 'weak',
+//   };
+// } else {
+//   gcTypes = {
+//     [k.NODE_PERFORMANCE_GC_MINOR]: 'minor',      // 1
+//     [k.NODE_PERFORMANCE_GC_MAJOR]: 'major',      // 2
+//     [k.NODE_PERFORMANCE_GC_INCREMENTAL]: 'incr', // 4
+//     [k.NODE_PERFORMANCE_GC_WEAKCB]: 'weak',      // 8
+//   };
+// }
+
+// //
+// // setup measurements with performance hooks
+// //
+// const verbose = process.env.VERBOSE;
+
+// const obs = new PerfObserver((list) => {
+//   const entries = list.getEntries();
+//   for (let i = 0; i < entries.length; i++) {
+//     const entry = entries[i];
+//     if (entry.entryType === 'measure') {
+//       verbose && console.log(`perf ${entry.name}: ${entry.duration}`);
+//       groupTimes.push(entry.duration);
+//     } else if (entry.entryType === 'gc') {
+//       if (verbose) {
+//         const {kind, flags} = entry.detail ? entry.detail : entry;
+//         console.log(`perf gc: ${entry.duration} (${gcTypes[kind]}) flags: ${flags}`);
+//       }
+//       gcCounts += 1;
+//       totalGCTime += entry.duration;
+//     }
+//   }
+// });
+// obs.observe({entryTypes: ['measure', 'gc'], buffered: true});
+
+// //
+// // function to run the setups, tests, and final.
+// //
+// async function test() {
+//   // call the tester's setup
+//   if (definitions.setup) {
+//     await (async() => definitions.setup(config))();
+//   }
+//   if (groupSetup) {
+//     await (async() => groupSetup(config))();
+//   }
+//   // warmup
+//   for (let i = warmupIterations; i > 0; i--) {
+//     await execute(functionChain);
+//   }
+//   await pause(groupWaitMS);
+
+//   // execute x groups of y iterations with a pause after each group
+//   for (let i = 0; i < groupCount; i++) {
+//     // setup for the group
+//     if (groupSetup) {
+//       await (async() => groupSetup(config))();
+//     }
+//     perf.mark('start-iteration');
+//     for (let i = groupIterations; i > 0; i--) {
+//       await execute(functionChain);
+//     }
+//     perf.measure('iteration-time', 'start-iteration');
+//     if (memCheck) {
+//       memory[i] = process.memoryUsage();
+//     }
+//     await pause(groupWaitMS);
+//   }
+
+//   // final pause - wait extra time to let GC finish.
+//   await pause(groupWaitMS * 10);
+//   return pause(groupWaitMS);
+// }
+
+
+// //
+// // execute functionChains
+// //
+// async function execute(fc) {
+//   let lastResult = undefined;
+//   for (let i = 0; i < fc.length; i++) {
+//     lastResult = await fc[i](lastResult);
+//   }
+// }
+
+// async function pause(ms) {
+//   return new Promise((resolve) => setTimeout(resolve, ms));
+// }
+
+// //
+// // execute the tests then report
+// //
+// test().then(() => {
+//   obs.disconnect();
+//   const gTimes = groupTimes.slice();
+//   const gcStats = {gcCounts, totalGCTime};
+//   const data = {gTimes, gcStats, stddevRange, config};
+//   if (memCheck) {
+//     data.memory = memory;
+//   }
+//   // if the test-definitions specifies a final execute it. it can
+//   // be output or cleanup or whatever.
+//   if (final) {
+//     final(config);
+//   }
+
+//   summarize(data, { json, terse });
+// });
